Improve timer system
Make the frontend file look cleaner and handle everything through
the backend chat.js file. Also add the ability to set a timer for
any string.

The only problem with this implemenentation is that
the words for the timer have to be listed in another function
otherwise the setInteval will not work. However, I will accomplish
this through creating a json file with responses that the user can
manually set and will persist through bot shutdown instead of having
a messy switch statement

Signed-off-by: kingbri <bdashore3@gmail.com>