bot: Make timers completely custom
The user can now add custom messages using the addtimermessage command.
Doing this stores it in RAM and it can be written to the JSON file through
timer write.

Also add protections against the user's syntax errors so they don't crash
the bot. If the user inputs the wrong syntax for timer start, the
bot tells the user to re-enter their syntax

Signed-off-by: kingbri <bdashore3@gmail.com>