king: Add a webhook to check if user's gone live
This webhook is started on bot connection and triggers ONLY
if the api realizes that the stream is live. Twitch API lags for
a minute or two before realizing that the user is live. Once this
event happens, the webhook causes something to happen.

Currently, I'm using ngrok for local testing, but this will change
once I migrate this to my VPS where the bot constantly runs

This something will be function calls to start timers so the user
doesn't have to enter the command every time the stream starts or
on any chat event.

Function calls will be implemented later down the road.

Signed-off-by: kingbri <bdashore3@gmail.com>