quotes: Add quotes functionality
Quoting is now possible within the bot.
To add a quote, it's *prefix*quote add and the quote is stored in RAM
To retrieve the quote, it's *prefix*quote retrieve and quote is pushed to chat
All quotes are written to a JSON file by *prefix*quote write
A quote can be remove by *prefix*quote remove

To make sure there are no overwrites, there is a check in the add
function which tells the user that the number is taken!

TODO: Add the ability to list all quotes in a user's twitch DMs

Signed-off-by: kingbri <bdashore3@gmail.com>