webhook: Adapt for server-side configuration
The ngrok was only a testing configuration and is now purged.
If you're using a webhook, it's a good idea to have a server, or
at least forward a port through your router instead of relying
on services like ngrok which can cost money if it's a permanent solution

Remove all instances of reverse proxy and hostname for the webhook
listener as it is not needed. The bot only has to listen on the
provided port (Assuming it's open)

Also clean up some formatting in chat.js

Signed-off-by: kingbri <bdashore3@gmail.com>