helpers: add baseline for lurk function
If the user isn't lurking, the command creates a new date and stores
the js time within an object called lurkTime. Al

If the username exists, then another date is created and the times
from the object date and the new date are subtracted and the difference
is in ms.

To convert from ms to hh:mm:ss, I used a function from stackoverflow
which takes the value and calculates everything in pretty format.

TODO: Make the chat responses prettier. Add the ability to stop
lurking. Add a list command?

Signed-off-by: kingbri <bdashore3@gmail.com>