lurk: Implement part 2
The command now shows the format in a neatly printed statement.
It converts the js time to actual time in hh:mm:ss.ms and the time
function stores everything in an array that is passed through the
lurk function and interpreted within the frontend.

Also add the ability to cancel lurks by removing the username from
the object and instantly return if the cancel argument is true.
tell the stream that the user is no longer lurking!

Signed-off-by: kingbri <bdashore3@gmail.com>