king: Add support for shouting out
if the user is in chat, the user will be shouted out. There is an
admin guard in place that deletes the message if a user tries to shoutout
without permission.

If the admin puts an @ by mistake when shouting out, the bot
will automatically remove the @ and shout out the username

Also, add new webhook settings

Signed-off-by: kingbri <bdashore3@gmail.com>