bot: initial commit
Create a kingbot base in C#.

Organized design where all information is loaded in its own class,
all commands are handled in its own class, and the client is exposed
to the entire project.

TODO: Write new commands along with methods to remove indicies
from the string list for a command.

Signed-off-by: kingbri <bdashore3@gmail.com>