botwide: Start structuring commands
Add a quotes command for the user to RETRIEVE quotes. This is
a Proof of Concept for working with the mongodb database that I
set up.

All individual command handling is stored in its own module.
It's more organized this way.

Also, make the commandhandler run asynchronously. This
will allocate commands on different threads making bot response
times faster.

HUGE thanks to tanish2k09 for helping me along the way.

Signed-off-by: kingbri <bdashore3@gmail.com>