Botwide: Make the entire thing async
Making the bot asynchronous allows for every task to be completed
on a seperate thread or queue, therefore improving overall performance.
This is the standard convention when making bots and I have finally
adopted it.

Doing this allows for commands which are mainly async in nature
to properly interact with the CommandHandler instead of dancing
between synchronous and asynchronous operations.

The only time when the operations are not Tasks is when we're
writing a single line to the console or doing something else
that doesn't require another thread.

A huge amount of this change was based/used off of code from
tanish2k09's bender bots which has an improved credentials helper
along with an asynchronous base. Therefore, he is co-authored in
this commit even though he is not part of the project.

Co-Authored-by: kingbri <bdashore3@gmail.com>
Co-Authored-by: Tanish Manku <tanish2k09.dev@gmail.com>