bot: Add intervals and custom commands
This is a massive update that is sectioned into 3 phases

1. Redo the commandhandler to pass the words into their respective
command modules and handle all subcommands in the module.

2. Add the ability for interval messages (messages that are posted
in chat over a certain interval). These intervals are customized
and stored in the mongodb collection "intervals". When the timer is
called, they're stored in a dictionary to be refrenced when the timer
needs to be stopped since stops are refrenced in a seperate function.

3. Add custom commands. Since we're using mongodb, we're doing something
different than reading from an object. I determined that checking if
the command exists in the database using the new ensure check rather
than reading all fields from the database into a dictionary. It's faster
and easier to program. I will eventually implement arguments since
C# is a lot more advanced and easier to work with than js.

Signed-off-by: kingbri <bdashore3@gmail.com>