bot: Migrate code to Aerospike database
Create a new helper and C# interface with the aerospike database.

Aerospike is a nosql database and stores data in a hybrid form
in ram and in disk. This database is MUCH faster and easier to work
with compared to mongodb.

Mongo in C# starts off easy, but eventually becomes extremely
difficult when trying to use commands with something other than
CRUD operations.

Additionally, the Aerospike async client does not abuse async/await
which makes overall code cleaner and easier to read. As a result,
the entire commandhandler has been rewritten with normal C# voids/types
rather than using Tasks.

The AsyncClient is completely thread safe. I have not implemented
the write and read handlers that Aerospike documentation uses
in their async guidelines, however CRUD works perfectly fine with
the AsyncClient.

Aerospike Documentation reference for C#:
https://www.aerospike.com/docs/client/csharp/

Signed-off-by: kingbri <bdashore3@gmail.com>