Add hooks for the API client
Adding twitch API support allows us to listen to the stream channel
and subscribe to events such as if the stream is live, if follows
are given, subs, hosts, etc. This function is meant to replace
streamlabs events in chat so we can use the '!' prefix again
without sacrificing goodies from streamlabs.

Also improve overall code style and comments for the production release

Signed-off-by: kingbri <bdashore3@gmail.com>