Botwide: Do a MASSIVE rewrite
This consists of four major parts
1. Remove the overuse of "static". Making things static basically
removes the actual use for classes as a whole. They should only
be used when declaring extended methods (aka making things global
on purpose) and NOT for convenience. This took a long time to purge
and was massively fixed due to part 4 of this commit.

2. Change to a postgresql database. I did this because postgresql is
free, open source, and extremely easy to configure using Entity
Framework Core (EFC) which PREVENTS Sql injection. The reason I am using
Postgres is because I don't want to use a single-file database like
SQLite and postgres is the fastest by far and guarantees ACID
due to the nature of Sql databases as a majority.

3. I ran into a problem when configuring Entity Framework Core and
this is hiding the connection string. On Runtime, this works properly,
but the problem was when I was creating a migration. On migration creation,
I couldn't execute the GetCreds method and therefore EFC only saw
"null". This was fixed by adding a Design-Time-Factory which gives
a set of steps to EFC when creating a model for the SQL database.
However, there was a problem when calling a use case in CRUD due to
the new "options" parameter. This was solved using part 4.

4. Add in Dependency Injection support. Dependency injection is exactly
what it sounds. If a function requires another function, dependency
injection can satisfy that requirement on runtime by using a simple
constructor from a service interface. I modified Program.cs to
"mark" the classes that would also be counted as services and this
removed the need for "static" overuse and is cleaner code practice
in general. Also, I created a service for the database context which
no longer forces me to hardcode the database context in a use statement.

TODO:
- Universalize CRUD operations (They are currently seperated by module
which is a big no)
- Add comments to the new changes
- Remove redundancies

HUGE thanks to the C# discord server. Without their constant help and support,
I wouldn't have learned this stuff easily. So give them some love as well.

Signed-off-by: kingbri <bdashore3@gmail.com>