Add in shoutout support and stream notifications
Shoutouts are only given by admins of the bot. If the person isn't
an admin, a shoutout isn't given.

TODO: Delete the person's message if he/she enters an admin command.

Add support for almost all events on twitch. Kingbot currently has
support for:
- Follows
- Hosts
- First-time subscribers
- Recurring subscribers
- Subscriber gifts
- ANONYMOUS subscriber gifts

This is a small, yet powerful change that allows people to move
away from using streamlabs cloudbot along with kingbot and for
kingbot to use the prefix '!' without conflicting with streamlabs
and streamelements.

Signed-off-by: kingbri <bdashore3@gmail.com>