botwide: Huge update
Add in lurk and list commands. These work with twitch whispers
to give the user a list of the avaliable choices per command.

The lurk command is made for lurkers! If a user wants to lurk,
he/she just enters !lurk and a clock starts. The user can periodically
enter the lurk command again to check how long he/she has been lurking
for.

If the user isn't lurking, just type !lurk cancel and the lurk
clock is cleared.

Signed-off-by: kingbri <bdashore3@gmail.com>