Fix FollowService and add ability for clearing
Dictionaries will continuously store items in memory until they're
forced to clear, or the bot service is stopped. To make this
process automatic, every time the stream stops, the dictionaries
"lurkTimes" and "intervals" are cleared. If an admin needs
to clear the dictionaries manually, they can execute the individual
command with clear after it like so with intervals:

?interval clear

In addition, FollowerService was spamming old follows on startup.
There is an issue regarding this and the way to fix it is outlined
in TwitchBot.cs. There should be a fix coming in the future, but
we don't know when that is.

Also, update the csproj to use the latest CI build of twitchlib client,
api, and pubsub since they're updated and deemed stable. The NuGet
pipeline's messed up and it's not showing the right builds.

Signed-off-by: kingbri <bdashore3@gmail.com>