Remove mut references
Instead of making the entire bot struct a mutable reference in a function's
arguments, I found that the writer only needs to be mut when sending messages.

Use interior mutability (a Mutex) on the bot struct's writer. Doing this
allows me to add the command map in the struct rather than introducing
it separately.

Also add a TypeMap using typemap_rev to store extra bot context data
along with a database connection for later use.

Signed-off-by: kingbri <bdashore3@gmail.com>