Add quote support
Quotes have been changed in v3.0 of KingBot.

You can add, remove, and list quotes using the given subcommands.
However, retrieval has been modified.

Quotes now function through aliases. For example, `quote add oof BIG OOOF`
will add a new quote under the alias of `oof` with the content `BIG OOOF`.

If you provide the quote command with a number/alias, it will retrieve
that quote.

Number references are now based on the position of the quote alias in the
database.

Also fix some extra things in regards to the rest of the bot for
fluidity and smoothness.

Signed-off-by: kingbri <bdashore3@gmail.com>