Add custom command support
Custom commands (such as !discord and !twitter) are important with
twitch bots. Add a configurable option to use these commands

The configuation can only be done by moderators of the channel,
but anyone can use them in the chat.

Hardcoded commands can't be set as custom ones due to alias conflicts

Signed-off-by: kingbri <bdashore3@gmail.com>