Add lurk command
When a user starts lurking, one can simply input the lurk command
and that starts a stopwatch.

The stopwatch keeps running until the stream ends and the lurker
can see the time by executing lurk again.

If the lurker wants to stop lurking, use lurk cancel to say that
the lurker's back.

If the mods want to clear the lurk map, clear removes all the entries.

Using a dashmap is ideal here since we are passing it between threads
and a mutexed hashmap is cumbersome

Signed-off-by: kingbri <bdashore3@gmail.com>