Add a CommandInfo struct
In every command, include some necessary information that may be used
inside each function.

The current two parts are the command length and a vector of the words. The reason
I used a Vector instead of an iterator is because a Vector is easier
to work with instead of an iterator.

More parts will be added as needed, but these are the only two fields
I need.

Signed-off-by: kingbri <bdashore3@gmail.com>