Add various features
Implement custom prefixes, the ability to connect to a database,
and a TypeMap that's passed through the Bot struct.

Doing these two things allow for the addition of quotes. Import
and format these from the 0.11 version.

This command stores quotes within a database. If the user wants to
retrieve said quote, he/she can either use the alias or the number given
by the list command.

Only moderators can add/remove quotes.

Signed-off-by: kingbri <bdashore3@gmail.com>