Add interval support
The updated version of the library made creating looping timers
easier.

When a timer is started, a task is spawned that includes a loop.
This loop keeps pushing a message to the channel every x seconds
and on each iteration, it checks whether the loop needs to break.

The breakage is controlled by a vector. If the vector contains both
the timer alias along with the channel name, that means the timer
has to be stopped.

Signed-off-by: kingbri <bdashore3@gmail.com>