Update all other commands
Add in lurk and custom commands.

Lurk is a unique feature to KingBot. If a user enters lurk once,
a timer would start. If the user enters lurk again, it shows how long
a user has lurked for.

Lurks can be cancelled by the user and removed by moderators.

Custom commands function in the same way as in CourtJester. However,
channel is replaced by the twitch channel and the user is replaced by
the user's username.

Signed-off-by: kingbri <bdashore3@gmail.com>