Add webhooks, shoutouts, and migrations
Webhooks: Twitch allows for bots to subscribe to webhooks where
information is sent in regards to follows and if the user
goes live. This will eventually be used to execute certain tasks
and let the stream know that their favorite creator is now live!

Shoutouts: Basically the `so` command in other bots, except
you can configure what message to say. If the user doesn't want
to configure a message, then a default message is provided.

Migrations: Allows for the database to be created with a single command.

Signed-off-by: kingbri <bdashore3@gmail.com>